Python-Projects/Hand-Gesture-ESP32-Light-Control/requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.11
pyserial==3.5
